+++ UamobiUtil/widgets/ElementWidgets/MegaIconButton.cpp
@@ -19,7 +19,7 @@ void MegaIconButton::paintEvent(QPaintEvent*)
 	spainter.drawPrimitive(QStyle::PE_Widget, opt);
 	opt.text = "";
 	spainter.drawControl(QStyle::CE_PushButton, opt);
-	spainter.drawItemPixmap(this->rect(), Qt::AlignCenter, this->ico.pixmap(this->rect().height() * imageCoefficient));
+    spainter.drawItemPixmap(this->rect(), Qt::AlignCenter, this->ico.pixmap(int(this->rect().height() * imageCoefficient)));
 	spainter.drawItemText(this->rect(), Qt::AlignBottom | Qt::AlignCenter, QPalette(), true, text(), QPalette::NoRole);
 }
 //	Changed default imgCoef 0.7 > 0.6 for better view
